keystone and glance
- keystone script now cascades return values from keystone setup
commands correctly to other commands
- beginning to work on troubleshooting glance

+++ folsom/openstack_keystone.sh
@@ -16,7 +16,9 @@ HOST_IP=$(/sbin/ifconfig eth0| sed -n 's/.*inet *addr:\([0-9\.]*\).*/\1/p')
 
 echo;
 echo "#############################################################################################################"
+echo;
 echo "The main IP address for this machine is probably $HOST_IP.  If that's wrong, ctrl-c and edit this script."
+echo;
 echo "#############################################################################################################"
 echo;
 
@@ -49,12 +51,14 @@ EOF
 . ./stackrc
 
 # edit keystone conf file to use templates and mysql
-cp /etc/keystone/keystone.conf /etc/keystone/keystone.conf.orig
+if [ -f /etc/keystone/keystone.conf.org ]; then
+  echo "Original backup of keystone.conf file exists. Your current config may be modified by this script."
+else
+  cp /etc/keystone/keystone.conf /etc/keystone/keystone.conf.orig
+fi
+
 sed -e "
 /^admin_token = ADMIN/s/^.*$/admin_token = $token/
-/^driver = keystone.catalog.backends.sql.Catalog/d
-/^\[catalog\]/a driver = keystone.catalog.backends.templated.TemplatedCatalog 
-/^\[catalog\]/a template_file = /etc/keystone/default_catalog.templates
 /^connection =.*$/s/^.*$/connection = mysql:\/\/keystone:$password@127.0.0.1\/keystone/
 " -i /etc/keystone/keystone.conf
 
@@ -88,41 +92,45 @@ keystone user-role-add --user-id $ADMIN_USER --role-id $ADMIN_ROLE --tenant-id $
 keystone user-role-add --user-id $ADMIN_USER --role-id $KEYSTONEADMIN_ROLE --tenant-id $ADMIN_TENANT
 keystone user-role-add --user-id $ADMIN_USER --role-id $KEYSTONESERVICE_ROLE --tenant-id $ADMIN_TENANT
 
+# the following commands use an interesting pattern due to keystone's asinine way of doing asset association.
+# we run a keystone command through the get_id function (defined above) and then use the resulting md5 hash  
+# to set a variable which we then use on the next command, effectively tying the two resources together inside
+# keystone.  we do this twice for each role.  why on earth keystone itself doesn't do this is anyone's guess.
+# consider me disgruntled.  - kord
+
 # nova
 NOVA_USER=$(get_id keystone user-create --name=nova --pass="$SERVICE_PASSWORD" --tenant-id $SERVICE_TENANT --email=$email)
 keystone user-role-add --tenant-id $SERVICE_TENANT --user-id $NOVA_USER --role-id $ADMIN_ROLE
-keystone service-create --name nova --type compute --description 'OpenStack Compute Service'
-keystone endpoint-create --region $KEYSTONE_REGION --service-id compute --publicurl 'http://'"$HOST_IP"':8774/v2/'"$NOVA_USER"'' --adminurl 'http://'"$HOST_IP"':8774/v2/'"$NOVA_USER"'' --internalurl 'http://'"$HOST_IP"':8774/v2/'"$NOVA_USER"''
-
-exit
+NOVA_ENDPOINT=$(get_id keystone service-create --name nova --type compute)
+keystone endpoint-create --region $KEYSTONE_REGION --service-id $NOVA_ENDPOINT --publicurl 'http://'"$HOST_IP"':8774/v2/%(tenant-id)s' --adminurl 'http://'"$HOST_IP"':8774/v2/%(tenant_id)s' --internalurl 'http://'"$HOST_IP"':8774/v2/%(tenant_id)s'
 
 # glance
 GLANCE_USER=$(get_id keystone user-create --name=glance --pass="$SERVICE_PASSWORD" --tenant_id $SERVICE_TENANT --email=$email)
 keystone user-role-add --tenant-id $SERVICE_TENANT --user-id $GLANCE_USER --role-id $ADMIN_ROLE
-keystone service-create --name glance --type image --description 'OpenStack Image Service'
-keystone endpoint-create --region $KEYSTONE_REGION --service-id glance --publicurl 'http://'"$HOST_IP"':9292/v2' --adminurl 'http://'"$HOST_IP"':9292/v2' --internalurl 'http://'"$HOST_IP"':9292/v2'
+GLANCE_ENDPOINT=$(get_id keystone service-create --name glance --type image)
+keystone endpoint-create --region $KEYSTONE_REGION --service-id $GLANCE_ENDPOINT --publicurl 'http://'"$HOST_IP"':9292/v2' --adminurl 'http://'"$HOST_IP"':9292/v2' --internalurl 'http://'"$HOST_IP"':9292/v2';
 
 # quantum
-if $SG_QUANTUM; then
-	QUANTUM_USER=$(get_id keystone user-create --name=quantum --pass="$SERVICE_PASSWORD" --tenant-id $SERVICE_TENANT --email=$email)
-	keystone user-role-add --tenant-id $SERVICE_TENANT --user-id $QUANTUM_USER --role-id $ADMIN_ROLE
-	keystone service-create --name quantum --type network --description 'OpenStack Networking service'
-	keystone endpoint-create --region $KEYSTONE_REGION --service-id quantum --publicurl 'http://'"$HOST_IP"':9696/' --adminurl 'http://'"$HOST_IP"':9696/' --internalurl 'http://'"$HOST_IP"':9696/'
+if [ "$SG_QUANTUM" != "0" ]; then
+  QUANTUM_USER=$(get_id keystone user-create --name=quantum --pass="$SERVICE_PASSWORD" --tenant-id $SERVICE_TENANT --email=$email)
+  keystone user-role-add --tenant-id $SERVICE_TENANT --user-id $QUANTUM_USER --role-id $ADMIN_ROLE
+  QUANTUM_ENDPOINT=$(get_id keystone service-create --name quantum --type network)
+  keystone endpoint-create --region $KEYSTONE_REGION --service-id $QUANTUM_ENDPOINT --publicurl 'http://'"$HOST_IP"':9696/' --adminurl 'http://'"$HOST_IP"':9696/' --internalurl 'http://'"$HOST_IP"':9696/'
 fi
 
 # cinder
 CINDER_USER=$(get_id keystone user-create --name=cinder --pass="$SERVICE_PASSWORD" --tenant-id $SERVICE_TENANT --email=$email)
 keystone user-role-add --tenant-id $SERVICE_TENANT --user-id $CINDER_USER --role-id $ADMIN_ROLE
-keystone service-create --name cinder --type volume --description 'OpenStack Volume Service'
-keystone endpoint-create --region $KEYSTONE_REGION --service-id volume --publicurl 'http://'"$HOST_IP"':8776/v1/$(tenant_id)s' --adminurl 'http://'"$HOST_IP"':8776/v1/$(tenant_id)s' --internalurl 'http://'"$HOST_IP"':8776/v1/$(tenant_id)s'
+CINDER_ENDPOINT=$(get_id keystone service-create --name cinder --type volume)
+keystone endpoint-create --region $KEYSTONE_REGION --service-id $CINDER_ENDPOINT --publicurl 'http://'"$HOST_IP"':8776/v1/$(tenant_id)s' --adminurl 'http://'"$HOST_IP"':8776/v1/$(tenant_id)s' --internalurl 'http://'"$HOST_IP"':8776/v1/$(tenant_id)s'
 
 # keystone 
-keystone service-create --name keystone --type identity --description 'OpenStack Identity'
-keystone endpoint-create --region $KEYSTONE_REGION --service-id identity --publicurl 'http://'"$HOST_IP"':5000/v2.0' --adminurl 'http://'"$HOST_IP"':35357/v2.0' --internalurl 'http://'"$HOST_IP"':5000/v2.0'
+KEYSTONE_ENDPOINT=$(get_id keystone service-create --name keystone --type identity) 
+keystone endpoint-create --region $KEYSTONE_REGION --service-id $KEYSTONE_ENDPOINT --publicurl 'http://'"$HOST_IP"':5000/v2.0' --adminurl 'http://'"$HOST_IP"':35357/v2.0' --internalurl 'http://'"$HOST_IP"':5000/v2.0'
 
 # ec2 compatability
-keystone service-create --name ec2 --type ec2 --description 'OpenStack EC2 service'
-keystone endpoint-create --region $KEYSTONE_REGION --service-id ec2 --publicurl 'http://'"$HOST_IP"':8773/services/Cloud' --adminurl 'http://'"$HOST_IP"':8773/services/Admin' --internalurl 'http://'"$HOST_IP"':8773/services/Cloud'
+EC2_ENDPOINT=$(get_id keystone service-create --name ec2 --type ec2)
+keystone endpoint-create --region $KEYSTONE_REGION --service-id $EC2_ENDPOINT --publicurl 'http://'"$HOST_IP"':8773/services/Cloud' --adminurl 'http://'"$HOST_IP"':8773/services/Admin' --internalurl 'http://'"$HOST_IP"':8773/services/Cloud'
 
 # create ec2 creds and parse the secret and access key returned
 RESULT=$(keystone ec2-credentials-create --tenant-id=$ADMIN_TENANT --user-id=$ADMIN_USER)
@@ -130,12 +138,16 @@ ADMIN_ACCESS=`echo "$RESULT" | grep access | awk '{print $4}'`
 ADMIN_SECRET=`echo "$RESULT" | grep secret | awk '{print $4}'`
 
 # write the secret and access to ec2rc
-cat > $EC2RC <<EOF
+cat > ec2rc <<EOF
 ADMIN_ACCESS=$ADMIN_ACCESS
 ADMIN_SECRET=$ADMIN_SECRET
 EOF
 
-echo "######################################################################################"
+echo "########################################################################################"
+echo;
+echo "Your EC2 credentials have been saved into ./ec2rc"
+echo;
 echo "Time to test keystone.  Do a '. ./stackrc' then a 'keystone user-list'."
 echo "Assuming you get a user list back, go on to install glance with './openstack_glance.sh'."
-echo "######################################################################################"
+echo;
+echo "########################################################################################"
